change  Task2 and Task3 from homework2

+++ src/homework/Tasks2With3DigitNumbers.java
@@ -5,11 +5,11 @@ public class Tasks2With3DigitNumbers {
         int number = 605;
         boolean t;
         System.out.println("number = " + number);
-        double hundreds = number / 100;
+        int hundreds = number / 100;
         System.out.println("hundreds = " + hundreds);
         double tens = (number / 10) % 10;
         System.out.println("tens = " + tens);
-        int unit = number % 10;
+        double unit = number % 10;
 
         if (unit == 0 ){
             System.err.println("Միավորի արժեքը պետք է լինի ոչ 0");
